Refactor task types, preliminary devcontainer
