+++ lib/task-types.ts
@@ -1,0 +1,22 @@
+
+export type TaskState = 'pending' | 'completed';
+
+export enum TaskPriority {
+    Unknown,
+    Critical,
+    High,
+    Medium,
+    Low,
+    Obsolete = 99
+}
+
+export interface NewTaskData extends Record<string, any> {
+    title: string;
+    priority: TaskPriority;
+}
+
+export interface Task extends Record<string, any> {
+    id: string;
+    title: string;
+    state: TaskState;
+}
